Moved api search logic to addflight file
