+++ client/src/components/dashboard/flights/addFlight.js
@@ -14,6 +14,8 @@ import MenuItem from "@mui/material/MenuItem";
 import FormControl from "@mui/material/FormControl";
 import InputLabel from "@mui/material/InputLabel";
 
+import moment from "moment";
+
 const AddFlight = ({
   flightData,
   tripId,
@@ -21,10 +23,7 @@ const AddFlight = ({
   open,
   userId,
   handleClose,
-  handleClear,
-  handleApiSearch
 }) => {
-
   const [flight, setFlight] = useState({
     flightNumber: flightData.flightNumber,
     departureTime: flightData.departureTime,
@@ -44,7 +43,6 @@ const AddFlight = ({
     trip: tripId,
   });
 
-
   const handleChange = (e) => {
     const value = e.target.value;
     setFlight({
@@ -55,7 +53,6 @@ const AddFlight = ({
 
   const onSubmit = (e) => {
     e.preventDefault();
-
     let url;
     if (flightId) {
       url = `http://localhost:8000/dashboard/flights/edit/${flightId}`
@@ -79,6 +76,7 @@ const AddFlight = ({
       bookingReference,
       isOutbound,
       user,
+      trip
     } = flight;
 
     const newFlight = {
@@ -96,31 +94,76 @@ const AddFlight = ({
       arrivalGate,
       bookingReference,
       isOutbound,
-      user,
-      trip: tripId,
+      user: userId,
+      trip
     };
 
     axios.post(url, newFlight).then((res) => {
       handleClose();
-      setFlight({
-        flightNumber: "",
-        departureTime: "",
-        departureDate: "",
-        airline: "",
-        departureAirport: "",
-        departureTerminal: "",
-        departureCity: "",
-        departureGate: "",
-        arrivalAirport: "",
-        arrivalTerminal: "",
-        arrivalCity: "",
-        arrivalGate: "",
-        bookingReference: "",
-        isOutbound: "",
-        user: userId,
-      })
+      handleClear();
     });
   };
+
+  const handleClear = () => {
+    setFlight({
+      flightNumber: "",
+      departureTime: "",
+      departureDate: "",
+      airline: "",
+      departureAirport: "",
+      departureTerminal: "",
+      departureCity: "",
+      departureGate: "",
+      arrivalAirport: "",
+      arrivalTerminal: "",
+      arrivalCity: "",
+      arrivalGate: "",
+      bookingReference: "",
+      isOutbound: ""
+    })
+  }
+
+  const formatDate = (date) => moment(date).format("YYYY-MM-DD");
+  const formatTime = (time) => moment(time).format("hh:mm");
+    const flightNumber = flight.flightNumber;
+    const flightDate = formatDate(flight.departureDate);
+    // console.log(flightNumber)
+    // console.log(flightDate)
+
+    const options = {
+      headers: {
+        'X-RapidAPI-Host': 'aerodatabox.p.rapidapi.com',
+        'X-RapidAPI-Key': process.env.REACT_APP_FLIGHT_API_KEY,
+      }
+    }
+    const flightApi = axios.create({
+      baseURL: 
+      `https://aerodatabox.p.rapidapi.com/flights/number/${flightNumber}/${flightDate}/`
+
+    });
+
+      const handleApiSearch = async () => {
+        await flightApi.get('/', options).then((res) => {
+          const data  = res.data[0]
+          console.log(res.data)
+  
+          setFlight({
+            ...flight, 
+            departureTime: formatTime(data.departure.scheduledTimeLocal),
+            airline: data.airline.name,
+            departureAirport: data.departure.airport.shortName,
+            departureTerminal: data.departure.terminal,
+            departureCity: data.departure.airport.municipalityName,
+            departureGate: data.departure.gate,
+            arrivalAirport: data.arrival.airport.name,
+            arrivalTerminal: data.arrival.terminal,
+            arrivalCity: data.arrival.airport.municipalityName,
+            arrivalGate: data.arrival.gate,
+            trip: tripId,
+            user: userId
+          })
+      })
+    }
 
   return (
     <div>
